moves gift item out and renames to explore
